Disabled download-images checkbox and worked with buttons in different states

+++ src/GiftCarder/app/src/main/java/ca/ualberta/smaccr/giftcarder/UserProfileActivity.java
@@ -61,7 +61,7 @@ public class UserProfileActivity extends Activity {
 
         Intent intent = getIntent();
         username = intent.getStringExtra(RegisterActivity.EXTRA_USERNAME);
-        profileState = (int) getIntent().getIntExtra(EXTRA_STATE, OWNER_STATE);
+        profileState = getIntent().getIntExtra(EXTRA_STATE, OWNER_STATE);
 
         User user = urc.getUser(username);
